# openapi/swagger/swagger.yaml
basePath: /
definitions:
...
  main.SwaggerEnvelope:
    properties:
      data: {}
      error:
        type: string
      pagination:
        $ref: '#/definitions/main.SwaggerPagination'
      success:
        type: boolean
    type: object
  main.SwaggerPagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  main.SwaggerRequest:
    additionalProperties: {}
    type: object
...
  model.ScheduleJobResp:
    properties:
      createdAt:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      expression:
        type: string
      id:
        type: string
      lastError:
        type: string
      lastRunAt:
        type: string
      lastStatus:
        type: string
      name:
        type: string
      runtime:
        $ref: '#/definitions/model.ScheduleRuntimeInfo'
      scheduleType:
        type: string
      taskKey:
        type: string
      updatedAt:
        type: string
    type: object
  model.ScheduleRuntimeInfo:
    properties:
      nextRun:
        type: string
      running:
        type: boolean
      scheduled:
        type: boolean
    type: object
...
  model.SystemSettingDefaultDTO:
    properties:
      category:
        type: string
      description:
        type: string
      key:
        type: string
      value:
        type: string
    type: object
  model.SystemSettingDefaultGroupsDTO:
    properties:
      alipay:
        items:
          $ref: '#/definitions/model.SystemSettingDefaultDTO'
        type: array
      announcement:
        items:
          $ref: '#/definitions/model.SystemSettingDefaultDTO'
        type: array
      basic:
        items:
          $ref: '#/definitions/model.SystemSettingDefaultDTO'
        type: array
      epay:
        items:
          $ref: '#/definitions/model.SystemSettingDefaultDTO'
        type: array
      wechat:
        items:
          $ref: '#/definitions/model.SystemSettingDefaultDTO'
        type: array
    type: object
  model.SystemSettingResp:
    properties:
      category:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      key:
        type: string
      updatedAt:
        type: string
      value:
        type: string
    type: object
  model.SystemSettingUpdateDTO:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  model.SystemSettingsDefaultsResp:
    properties:
      general:
        items:
          $ref: '#/definitions/model.SystemSettingDefaultDTO'
        type: array
      oauth:
        $ref: '#/definitions/model.SystemSettingDefaultGroupsDTO'
      operation:
        $ref: '#/definitions/model.SystemSettingDefaultGroupsDTO'
      payment:
        $ref: '#/definitions/model.SystemSettingDefaultGroupsDTO'
    type: object
  model.SystemSettingsUpdateReq:
    properties:
      settings:
        items:
          $ref: '#/definitions/model.SystemSettingUpdateDTO'
        type: array
    type: object
...
  schedule.TaskDescriptor:
    properties:
      description:
        type: string
      key:
        type: string
    type: object
...
info:
  contact: {}
  description: Contract for the Nex HTTP API.
  license:
    name: MIT
    url: https://opensource.org/license/mit
  title: Nex API
  version: 1.0.0
paths:
...
  /api/scheduled-jobs:
    get:
      operationId: scheduled_jobs_route_get
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ScheduleJobResp'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: GET /api/scheduled-jobs
      tags:
      - schedule
    post:
      consumes:
      - application/json
      operationId: scheduled_jobs_route_post
      parameters:
      - description: JSON request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.SwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.ScheduleJobResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: POST /api/scheduled-jobs
      tags:
      - schedule
  /api/scheduled-jobs/{id}:
    delete:
      operationId: scheduled_jobs_id_route_delete
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: DELETE /api/scheduled-jobs/{id}
      tags:
      - schedule
    get:
      operationId: scheduled_jobs_id_route_get
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.ScheduleJobResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: GET /api/scheduled-jobs/{id}
      tags:
      - schedule
    put:
      consumes:
      - application/json
      operationId: scheduled_jobs_id_route_put
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: JSON request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.SwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.ScheduleJobResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: PUT /api/scheduled-jobs/{id}
      tags:
      - schedule
  /api/scheduled-jobs/{id}/run:
    post:
      consumes:
      - application/json
      operationId: scheduled_jobs_id_run_route_post
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: POST /api/scheduled-jobs/{id}/run
      tags:
      - schedule
  /api/scheduled-jobs/tasks:
    get:
      operationId: scheduled_jobs_tasks_route_get
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schedule.TaskDescriptor'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: GET /api/scheduled-jobs/tasks
      tags:
      - schedule
...
  /api/system-settings:
    get:
      operationId: system_settings_route_get
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.SystemSettingResp'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: GET /api/system-settings
      tags:
      - system-settings
    put:
      consumes:
      - application/json
      operationId: system_settings_route_put
      parameters:
      - description: request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SystemSettingsUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: PUT /api/system-settings
      tags:
      - system-settings
  /api/system-settings/announcement:
    get:
      operationId: system_settings_announcement_route_get
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      summary: GET /api/system-settings/announcement
      tags:
      - system-settings
  /api/system-settings/defaults:
    get:
      operationId: system_settings_defaults_route_get
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.SystemSettingsDefaultsResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      summary: GET /api/system-settings/defaults
      tags:
      - system-settings
...
securityDefinitions:
  ApiTokenAuth:
    in: header
    name: Authorization
    type: apiKey
  CronSecretAuth:
    in: header
    name: Authorization
    type: apiKey
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"